// src/helpers/enums.ts
export enum CountryOptions {
    bra = "brazil",
    spa = "spain",
    usa = "usa",
    ita = "italy",
    ger = "germany",
    fra = "france",
    uk = "uk",
    ira = "iran",
    ned = "netherlands",
    bel = "belgium",
}

export enum LabelOptions {
    bra = "Brasil",
    spa = "Espanha",
    usa = "Estados Unidos",
    ita = "Itália",
    ger = "Alemanha",
    fra = "França",
    uk = "Reino Unido",
    ira = "Iran",
    ned = "Países Baixos",
    bel = "Bélgica",
}